+++ src/components/Footer.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+const Footer = () => (
+  <footer className="p-4 bg-blue-500 text-white text-center">
+    © 2025 Event Scheduler
+  </footer>
+);
+export default Footer;
