+++ src/components/Header.jsx
@@ -4,24 +4,35 @@ import { useAuth } from "../context/AuthProvider";
 import { useAppContext } from "../context/AppContext";
 
 const Header = () => {
-  const { isAuthenticated, login, logout } = useAuth();
   const { setCurrentView } = useAppContext();
+  const { isAuthenticated, logout } = useAuth();
 
   return (
-    <header className="p-4 bg-blue-600 text-white flex justify-between items-center">
-      <h1 className="text-2xl font-bold">Event App</h1>
-      <nav>
+    <header className="p-4 bg-blue-500 text-white flex justify-between items-center">
+      <img
+        src="src\asserts\logo_event_scheduler.png"
+        alt="Logo"
+        className="w-10 h-10 "
+      />
+      <h1 className="text-3xl font-bold">Event Scheduler</h1>
+      <div>
+        <button
+          onClick={() => setCurrentView("events")}
+          className="px-4 py-2  border border-white bg-transparent  rounded-lg mr-2"
+        >
+          Home
+        </button>
         {!isAuthenticated ? (
           <>
             <button
               onClick={() => setCurrentView("login")}
-              className="bg-gray-800 px-4 py-2 rounded mr-2"
+              className="px-4 py-2 bg-transparent  border border-white rounded-lg mr-2"
             >
               Login
             </button>
             <button
               onClick={() => setCurrentView("signup")}
-              className="bg-gray-800 px-4 py-2 rounded"
+              className="px-4 py-2  border border-white bg-transparent  rounded-lg"
             >
               Sign Up
             </button>
@@ -29,17 +40,20 @@ const Header = () => {
         ) : (
           <>
             <button
-              onClick={() => setCurrentView("events")}
-              className="bg-gray-800 px-4 py-2 rounded mr-2"
+              onClick={() => setCurrentView("create-event")}
+              className="px-4 py-2   border border-white bg-transparent rounded-lg mr-2"
             >
-              Events
+              Create Event
             </button>
-            <button onClick={logout} className="bg-red-500 px-4 py-2 rounded">
+            <button
+              onClick={logout}
+              className="px-4 py-2   border border-white bg-transparent rounded-lg"
+            >
               Logout
             </button>
           </>
         )}
-      </nav>
+      </div>
     </header>
   );
 };
